Allow force flush of oled display.
API copmatible refactoring of the oled_render function
that allows callers to forcibly render all dirty
blocks instead of allowing the oled_task function to
render them asynchronously. This is mostly useful
when want to display a logo/message/graphic right
before rebooting into a bootloader or other state
where the MCU will be unable to finish flushing blocks
to the display.