Adds a way to separate tab from AUTO_SHIFT_SPECIAL.
The goal is to add a means of enabling all of the other
special characters while disabling Auto Shift on tab.
While I was at it, this adds documentation to the predefined
character groups used to setup `get_auto_shifted_key` so
that you don't have to go digging through .h files for
them.